+++ square/api/catalog_api.py
@@ -197,15 +197,14 @@ class CatalogApi(BaseApi):
                              image_file=None):
         """Does a POST request to /v2/catalog/images.
 
-        Upload an image file to create a new
-        [CatalogImage](#type-catalogimage) for an existing
-        [CatalogObject](#type-catalogobject). Images can be uploaded and
-        linked in this request or created independently
-        (without an object assignment) and linked to a
-        [CatalogObject](#type-catalogobject) at a later time.
-        CreateCatalogImage accepts HTTP multipart/form-data requests with a
-        JSON part and an image file part in
-        JPEG, PJPEG, PNG, or GIF format. The maximum file size is 15MB. 
+        Uploads an image file to be represented by an
+        [CatalogImage](#type-catalogimage) object linked to an existing
+        [CatalogObject](#type-catalogobject) instance. A call to this endpoint
+        can upload an image, link an image to
+        a catalog object, or do both.
+        This `CreateCatalogImage` endpoint accepts HTTP multipart/form-data
+        requests with a JSON part and an image file part in
+        JPEG, PJPEG, PNG, or GIF format. The maximum file size is 15MB.
         Additional information and an example cURL request can be found in the
         [Create a Catalog Image
         recipe](https://developer.squareup.com/docs/more-apis/catalog/cookbook/
@@ -267,8 +266,9 @@ class CatalogApi(BaseApi):
     def catalog_info(self):
         """Does a GET request to /v2/catalog/info.
 
-        Returns information about the Square Catalog API, such as batch size
-        limits for `BatchUpsertCatalogObjects`.
+        Retrieves information about the Square Catalog API, such as batch
+        size
+        limits that can be used by the `BatchUpsertCatalogObjects` endpoint.
 
         Returns:
             CatalogInfoResponse: Response from the API. Success
@@ -565,22 +565,24 @@ class CatalogApi(BaseApi):
                                body):
         """Does a POST request to /v2/catalog/search.
 
-        Queries the targeted catalog using a variety of query expressions.
-        Supported query expressions are of the following types:
-        - [CatalogQuerySortedAttribute](#type-catalogquerysortedattribute),
-        - [CatalogQueryExact](#type-catalogqueryexact),
-        - [CatalogQueryRange](#type-catalogqueryrange),
-        - [CatalogQueryText](#type-catalogquerytext),
-        - [CatalogQueryItemsForTax](#type-catalogqueryitemsfortax),
-        -
-        [CatalogQueryItemsForModifierList](#type-catalogqueryitemsformodifierli
-        st),
-        -
-        [CatalogQueryItemsForItemOptions](#type-catalogqueryitemsforitemoptions
-        ), and
-        -
-        [CatalogQueryItemVariationsForItemOptionValues](#type-catalogqueryitemv
-        ariationsforitemoptionvalues).
+        Searches for [CatalogObject](#type-CatalogObject) of any types against
+        supported search attribute values, 
+        excluding custom attribute values on items or item variations, against
+        one or more of the specified query expressions, 
+        This (`SearchCatalogObjects`) endpoint differs from the
+        [SearchCatalogItems](#endpoint-Catalog-SearchCatalogItems)
+        endpoint in the following aspects:
+        - `SearchCatalogItems` can only search for items or item variations,
+        whereas `SearchCatalogObjects` can search for any type of catalog
+        objects.
+        - `SearchCatalogItems` supports the custom attribute query filters to
+        return items or item variations that contain custom attribute values,
+        where `SearchCatalogObjects` does not.
+        - `SearchCatalogItems` does not support the `include_deleted_objects`
+        filter to search for deleted items or item variations, whereas
+        `SearchCatalogObjects` does.
+        - The both endpoints have different call conventions, including the
+        query filter formats.
 
         Args:
             body (SearchCatalogObjectsRequest): An object containing the
@@ -600,6 +602,70 @@ class CatalogApi(BaseApi):
 
         # Prepare query URL
         _url_path = '/v2/catalog/search'
+        _query_builder = self.config.get_base_uri()
+        _query_builder += _url_path
+        _query_url = APIHelper.clean_url(_query_builder)
+
+        # Prepare headers
+        _headers = {
+            'accept': 'application/json',
+            'content-type': 'application/json; charset=utf-8'
+        }
+
+        # Prepare and execute request
+        _request = self.config.http_client.post(_query_url, headers=_headers, parameters=APIHelper.json_serialize(body))
+        OAuth2.apply(self.config, _request)
+        _response = self.execute_request(_request)
+
+        decoded = APIHelper.json_deserialize(_response.text)
+        if type(decoded) is dict:
+            _errors = decoded.get('errors')
+        else:
+            _errors = None
+        _result = ApiResponse(_response, body=decoded, errors=_errors)
+        return _result
+
+    def search_catalog_items(self,
+                             body):
+        """Does a POST request to /v2/catalog/search-catalog-items.
+
+        Searches for catalog items or item variations by matching supported
+        search attribute values, including
+        custom attribute values, against one or more of the specified query
+        expressions, 
+        This (`SearchCatalogItems`) endpoint differs from the
+        [SearchCatalogObjects](#endpoint-Catalog-SearchCatalogObjects)
+        endpoint in the following aspects:
+        - `SearchCatalogItems` can only search for items or item variations,
+        whereas `SearchCatalogObjects` can search for any type of catalog
+        objects.
+        - `SearchCatalogItems` supports the custom attribute query filters to
+        return items or item variations that contain custom attribute values,
+        where `SearchCatalogObjects` does not.
+        - `SearchCatalogItems` does not support the `include_deleted_objects`
+        filter to search for deleted items or item variations, whereas
+        `SearchCatalogObjects` does.
+        - The both endpoints use different call conventions, including the
+        query filter formats.
+
+        Args:
+            body (SearchCatalogItemsRequest): An object containing the fields
+                to POST for the request.  See the corresponding object
+                definition for field details.
+
+        Returns:
+            SearchCatalogItemsResponse: Response from the API. Success
+
+        Raises:
+            APIException: When an error occurs while fetching the data from
+                the remote API. This exception includes the HTTP Response
+                code, an error message, and the HTTP body that was received in
+                the request.
+
+        """
+
+        # Prepare query URL
+        _url_path = '/v2/catalog/search-catalog-items'
         _query_builder = self.config.get_base_uri()
         _query_builder += _url_path
         _query_url = APIHelper.clean_url(_query_builder)